// src/generated/graphql.tsx
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  timestamptz: any;
  uuid: any;
};
// ...
export type Comments = {
  __typename?: "comments";
  comment: Scalars["String"];
  created_at?: Maybe<Scalars["timestamptz"]>;
  id: Scalars["uuid"];
  /** An object relationship */
  todo?: Maybe<Todo>;
  todo_id?: Maybe<Scalars["uuid"]>;
  updated_at?: Maybe<Scalars["timestamptz"]>;
};
// ...
export type Todo = {
  __typename?: "todo";
  /** An array relationship */
  comments?: Array<Comments>;
  complete?: Maybe<Scalars["Boolean"]>;
  created_at?: Scalars["timestamptz"];
  description?: Maybe<Scalars["String"]>;
  id?: Scalars["uuid"];
  title?: Scalars["String"];
  updated_at?: Scalars["timestamptz"];
};